Trying to kill a few annoying warnings.

+++ src/gui/Source/Window Controllers/WindowControllerNumberMatrix.mm
@@ -90,7 +90,7 @@
     
     RevBayesCore::Trace* t = data->at(0);
     
-    return t->size();
+    return (int)t->size();
 }
 
 - (void)setControlWindowSize {
